+++ src/components/TextInput/TextInputTypes.tsx
@@ -23,6 +23,12 @@ export enum TextInputFormats {
   url = "https://domain.com",
 }
 
+export enum TextInputVariants {
+  Default = "default",
+  SearchBar = "searchBar",
+  SearchBarSelect = "searchBarSelect",
+}
+
 /*
 export enum TextInputPatterns {
   // eslint-disable-next-line prettier/prettier