+++ src/theme/components/button.ts
@@ -10,6 +10,10 @@ const baseStyle = {
   px: "s",
   textDecoration: "none",
   wordWrap: "normal",
+  svg: {
+    fill: "currentColor",
+    marginTop: "xxs",
+  },
   _hover: {
     bg: "ui.link.secondary",
   },
@@ -91,12 +95,16 @@ const callout = {
     bg: "brand.secondary",
   },
 };
+const searchBar = {
+  ...primary,
+  borderLeftRadius: "0",
+};
 
 const Button = {
   baseStyle,
   sizes,
   // Available variants:
-  // primary, secondary, link, pill, icon-only
+  // primary, secondary, link, pill, icon-only, searchBar
   variants: {
     primary,
     secondary,
@@ -104,6 +112,7 @@ const Button = {
     pill,
     ["icon-only"]: iconOnly,
     callout,
+    searchBar,
   },
   // Default values
   defaultProps: {