+++ src/theme/components/tabs.ts
@@ -18,7 +18,7 @@ const tab = {
   whiteSpace: "nowrap",
   _hover: {
     bg: "ui.gray.x-light-warm",
-    borderTopRadius: "2",
+    borderTopRadius: "sm",
     borderBottom: "1px solid",
     borderBottomColor: "ui.black",
   },
@@ -26,7 +26,7 @@ const tab = {
     fontWeight: "bold",
     bg: "ui.gray.light-warm",
     border: "0",
-    borderTopRadius: "2",
+    borderTopRadius: "sm",
     borderBottom: "3px solid",
     borderBottomColor: "ui.black",
     paddingBottom: "5px",