+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Exports Chakra's `Grid` and `GridItem` components.
 - Composes the DS `Label` component with Chakra component and updates the font size.
 - Composes the DS `StatusBadge` component from Chakra's component.
+- Composes the DS `Forms` component using DS Chakra based components.
 
 ### Changes